+++ src/main/java/org/parseplatform/client/ParseObject.java
@@ -18,9 +18,12 @@ package org.parseplatform.client;
 
 import com.dotweblabs.shape.client.Shape;
 import com.google.common.base.Joiner;
+import com.google.gwt.core.client.GWT;
 import com.google.gwt.json.client.*;
+import com.google.gwt.reflect.shared.GwtReflect;
 import com.google.gwt.user.client.rpc.AsyncCallback;
-import org.parseplatform.client.util.JSON;
+import elemental.client.Browser;
+//import org.parseplatform.client.util.JSON;
 
 import java.util.Arrays;
 import java.util.Iterator;
@@ -191,7 +194,7 @@ public class ParseObject extends JSONObject {
     public ParseACL getACL() {
         if(get("ACL") != null && get("ACL").isObject() != null) {
             JSONObject acl = get("ACL").isObject();
-            ParseACL parseACL = (ParseACL) acl;
+            ParseACL parseACL = new ParseACL(acl);
             return parseACL;
         }
         return null;
@@ -224,7 +227,7 @@ public class ParseObject extends JSONObject {
     }
 
     public Long getLong(String key) {
-        if(this.get(key) != null && this.get(key).isString() != null) {
+        if(this.get(key) != null && this.get(key).isNumber() != null) {
             Double db = this.get(key).isNumber().doubleValue();
             return Long.valueOf(db + "");
         }
@@ -234,6 +237,22 @@ public class ParseObject extends JSONObject {
     public Double getDouble(String key) {
         if(this.get(key) != null && this.get(key).isNumber() != null) {
             Double db = this.get(key).isNumber().doubleValue();
+            return db;
+        }
+        return null;
+    }
+
+    public Short getShort(String key) {
+        if(this.get(key) != null && this.get(key).isNumber() != null) {
+            Short db = new Double(this.get(key).isNumber().doubleValue()).shortValue();
+            return db;
+        }
+        return null;
+    }
+
+    public Float getFloat(String key) {
+        if(this.get(key) != null && this.get(key).isNumber() != null) {
+            Float db = new Double(this.get(key).isNumber().doubleValue()).floatValue();
             return db;
         }
         return null;
@@ -311,23 +330,41 @@ public class ParseObject extends JSONObject {
      * @param <T>
      * @return a new new {@ParseObject}
      */
-    public static <T> ParseObject from(String className, T object) {
-        String json = JSON.stringify(object);
-        ParseObject parseObject = ParseObject.parse(className, json);
-        return parseObject;
-    }
+//    public static <T> ParseObject from(String className, T object) {
+//        String json = JSON.stringify(object);
+//        ParseObject parseObject = ParseObject.parse(className, json);
+//        return parseObject;
+//    }
 
     /**
-     * Marshall this {@ParseObject} into a target object.
+     * Unmarshall this {@ParseObject} into a target object.
      *
      * @param clazz Target object class
      * @param <T> Target object type
      * @return Target object
      */
-    public <T> T as(Class<T> clazz) {
-        T as = JSON.parse(this.toString());
+    public <T> T unmarshall(Class<T> clazz) {
+        T as = null;
+        GwtUnmarshaller unmarshaller = GWT.create(GwtUnmarshaller.class);
+        try {
+            T instance = clazz.newInstance();
+            as = unmarshaller.unmarshall(clazz, instance, this);
+        } catch (Exception e) {
+            Browser.getWindow().getConsole().log(e.getMessage());
+        }
         return as;
     }
+
+    public static <T> ParseObject marshall(Class<T> clazz, Object instance) {
+        if(instance != null) {
+            GwtMarshaller marshaller = GWT.create(GwtMarshaller.class);
+            ParseObject parseObject = marshaller.marshall(instance);
+            return parseObject;
+        }
+        return null;
+    }
+
+
 
     /**
      * Check if this object is an error.